# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI智能助手系统",
    "summary": "AI智能助手系统API",
    "description": "\n## AI智能助手系统后端API\n\n基于FastAPI和LangChain 1.0框架构建的企业级AI智能助手系统，提供以下核心功能：\n\n### 核心功能\n\n* **智能对话** - 支持多轮对话、流式响应、对话历史管理\n* **RAG知识库** - 文档上传、向量化、智能检索问答\n* **Agent智能代理** - 工具调用、任务执行、推理过程可视化\n* **用户管理** - 注册登录、JWT认证、配额管理\n* **系统监控** - 使用统计、健康检查、Prometheus指标\n\n### 技术栈\n\n* **Web框架**: FastAPI 0.104+\n* **AI框架**: LangChain 1.0\n* **数据库**: MySQL 8.0, Redis 7.0, Chroma\n* **LLM**: 通义千问 (Tongyi Qianwen)\n\n### 认证方式\n\n大部分API端点需要JWT认证。在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 速率限制\n\n* 登录接口: 5次/分钟\n* 普通API: 100次/分钟\n* LLM调用: 20次/分钟\n\n### 错误响应格式\n\n所有错误响应遵循统一格式：\n```json\n{\n    \"error_code\": \"1001\",\n    \"message\": \"错误描述\",\n    \"request_id\": \"uuid\"\n}\n```\n    ",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge-bases": {
      "post": {
        "tags": [
          "知识库管理"
        ],
        "summary": "创建知识库",
        "description": "创建新的知识库，用于存储和管理文档。",
        "operationId": "create_knowledge_base_api_v1_knowledge_bases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "知识库管理"
        ],
        "summary": "获取知识库列表",
        "description": "获取当前用户的所有知识库列表，支持分页。",
        "operationId": "get_knowledge_bases_api_v1_knowledge_bases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseListResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{kb_id}": {
      "get": {
        "tags": [
          "知识库管理"
        ],
        "summary": "获取知识库详情",
        "description": "获取指定知识库的详细信息。",
        "operationId": "get_knowledge_base_api_v1_knowledge_bases__kb_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "知识库管理"
        ],
        "summary": "更新知识库",
        "description": "更新指定知识库的信息。",
        "operationId": "update_knowledge_base_api_v1_knowledge_bases__kb_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Kb Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "知识库管理"
        ],
        "summary": "删除知识库",
        "description": "删除指定知识库及其所有文档和向量数据。",
        "operationId": "delete_knowledge_base_api_v1_knowledge_bases__kb_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeBaseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "知识库名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "知识库描述"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "知识库分类"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreate",
        "description": "创建知识库请求"
      },
      "KnowledgeBaseListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "总数"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeBaseResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "知识库列表"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "KnowledgeBaseListResponse",
        "description": "知识库列表响应"
      },
      "KnowledgeBaseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "知识库ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "知识库名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "知识库描述"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "知识库分类"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count",
            "description": "文档数量",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeBaseResponse",
        "description": "知识库响应"
      },
      "KnowledgeBaseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "知识库名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "知识库描述"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "知识库分类"
          }
        },
        "type": "object",
        "title": "KnowledgeBaseUpdate",
        "description": "更新知识库请求"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "消息内容"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "通用消息响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}